ShortUrlController 개선
- ShortUrlController에 URL 단축 기능 맵핑
- TestCode 개선

work items: #9

+++ src/test/kotlin/com/tommy/urlshortener/service/ShortenKeyGeneratorTest.kt
@@ -34,7 +34,7 @@ class ShortenKeyGeneratorTest(
         val actual = sut.generate(timestamp)
 
         // Assert
-        assertThat(actual).isEqualTo("${timestamp}0300")
+        assertThat(actual).isEqualTo("${timestamp}0300".toLong())
     }
 
     @Test
